refactor(react-google-maps): directly import assets
